update hotel settings

+++ src/services/apiCabins.js
@@ -9,24 +9,23 @@ export async function getCabins() {
   }
   return data;
 }
-export async function createEditCabin(newCabin,id) {
-  const hasImagePath = newCabin.image?.startWith?.(supabaseUrl);
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll("/","");
-  const imagePath =hasImagePath ? newCabin.image :  `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    ""
+  );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-
-  
   // 1. create cabin
   let query = supabase.from("cabins");
   // A) CREATE
-  if(!id) query =
-   query
-    .insert([{...newCabin, image: imagePath}]);
-    // B) EDIT
-    if(id) query =
-      query.update({...newCabin, image: imagePath})
-    .eq('id', id);
-    const { data, error } = await query;
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+  const { data, error } = await query;
 
   if (error) {
     console.error(error);
@@ -34,16 +33,21 @@ export async function createEditCabin(newCabin,id) {
   }
 
   // 2. upload image
-  const { error: storageError } = await supabase
-  .storage
-  .from('cabin-images')
-  .upload(imageName, newCabin.image);
- // 3. Delete the cabin if there was en error uploading image
- if(storageError){
-  await supabase.from("cabins").delete().eq("id", data.id);
-  console.error(storageError);
-    throw new Error("cabin image could not be uploaded and the cabin was not created");
- }
+
+  if (hasImagePath) return data;
+  console.log("Uploading Image...");
+
+  const { error: storageError } = await supabase.storage
+    .from("cabin-images")
+    .upload(imageName, newCabin.image);
+  // 3. Delete the cabin if there was en error uploading image
+  if (storageError) {
+    await supabase.from("cabins").delete().eq("id", data.id);
+    console.error(storageError);
+    throw new Error(
+      "cabin image could not be uploaded and the cabin was not created"
+    );
+  }
   return data;
 }
